requirements.txt
openai
python-dotenv
mysql-connector-python
pandas
azure-search-documents
azure-core
streamlit